+++ eslint.config.js
@@ -3,13 +3,13 @@ import globals from "globals";
 import tseslint from "typescript-eslint";
 import { defineConfig, globalIgnores } from "eslint/config";
 export default defineConfig([
-    globalIgnores(["dist"]),
-    {
-        files: ["**/*.ts"],
-        extends: [js.configs.recommended, tseslint.configs.recommended],
-        languageOptions: {
-            ecmaVersion: 2020,
-            globals: globals.node,
-        },
+  globalIgnores(["dist"]),
+  {
+    files: ["**/*.ts"],
+    extends: [js.configs.recommended, tseslint.configs.recommended],
+    languageOptions: {
+      ecmaVersion: 2020,
+      globals: globals.node,
     },
+  },
 ]);