+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Create initial DB migration
 - Add more documentation including GPL license and code of conduct
+- Update AWS SDK and various other packages
+- Uninstall `@vercel/postgres`
+- `camelCase` middleware: Fix: `null` was being converted to `{}`
+
 
 ## 0.2.1 - 02/02/2026 13:30 UTC+1