+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Uninstall `@vercel/postgres`
 - `camelCase` middleware: Fix: `null` was being converted to `{}`
 - Add `duration` to scores
+- Restructure `personalBest` property to expose `outcome` directly in the `POST /login` response
 
 ## 0.2.1 - 02/02/2026 13:30 UTC+1