+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `camelCase` middleware: Fix: `null` was being converted to `{}`
 - Add `duration` to scores
 - Restructure `personalBest` property to expose `outcome` directly in the `POST /login` response
+- Update integration tests
 
 ## 0.2.1 - 02/02/2026 13:30 UTC+1