+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.3.0 -
+## 0.3.0 - 05/02/2026 19:00 UTC+1
 
 - Create initial DB migration
 - Add more documentation including GPL license and code of conduct
@@ -10,7 +10,7 @@
 - Add `duration` to scores
 - Restructure `personalBest` property to expose `outcome` directly in the `POST /login` response
 - Update integration tests
-- `PUT /player`: Change `dateOfBirth` into optional
+- `PUT /player`: Change `dateOfBirth` to optional and add `audioVolume` 
 
 ## 0.2.1 - 02/02/2026 13:30 UTC+1